added order and order details functionality as far as ordering

+++ DataAccess/Entities/OrderDetails.cs
@@ -10,9 +10,9 @@ namespace DataAccess
     {
         public int OrderID { get; set; }
         public int ProductID { get; set; }
-        public Decimal UnitPrice { get; set; }
+        public double UnitPrice { get; set; }
         public int Quantity { get; set; }
-        public Double Discount { get; set; }
+        public double Discount { get; set; }
 
         // reference 
         public Order Order { get; set; }
